scripts/ets_scrape/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
urllib3==2.0.7
openai>=1.50.0
PyMuPDF==1.23.8
python-dotenv==1.0.0
Pillow==10.1.0